requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Mail
itsdangerous
Flask-Bcrypt